✨: Redux의 slice를 관리하기 위해 디렉토리 생성 및 위치 변경, slice 네이밍 변경 예정, ManagerBar 비활성화시 view보드로 상태 변경 코드 추가 필요
