+++ shared/store/slice/factoryModeSlice.ts
