+++ shared/store/slice/index.ts
@@ -1,0 +1,2 @@
+export { default as testSlice } from './testSlice';
+export { default as modeReducer } from './factoryModeSlice';
