+++ shared/store/slice/testSlice.ts
@@ -1,0 +1,15 @@
+import { createSlice } from '@reduxjs/toolkit';
+const counterSlice = createSlice({
+  name: 'counter',
+  initialState: { count: 0 },
+  reducers: {
+    Increment: (state) => {
+      state.count += 1;
+    },
+    Decrement: (state) => {
+      state.count -= 1;
+    },
+  },
+});
+export const { Increment, Decrement } = counterSlice.actions;
+export default counterSlice.reducer;
